app/requirements.txt
ultralytics>=8.0.0
fastapi
uvicorn
python-multipart
pillow
pytest
httpx
python-decouple
